Switch to use GRUB EFI network boot for EFI client
New file set-default-grub-efi-img was added. It will be used
to set the menuentry for GRUB EFI NB was added.
New file drbl-gen-grub-efi-nb was added so that it can prepare
GRUB EFI network boot files (/tftpboot/nbi_img/bootx64.efi).
New file gen-grub-efi-nb-menu was added so that it can be used
to create the GRUB EFI NB config files in
/tftpboot/nbi_img/grub-efi.cfg/.
This is the initial version for using GRUB EFI NB as the uEFI
network boot clients due to the syslinux.efi is not working with iPXE clients.