Move variable secure_boot to drbl.conf.
Make variable secure_boot to be global, hence it's moved to drbl.conf
from drbl-gen-grub-efi-nb. Hence both drbl-gen-grub-efi-nb and drblpush
will use the same setting. By default secure_boot is not enabled due to
this issue for secure grub boot loader:
https://lists.gnu.org/archive/html/grub-devel/2016-04/msg00051.html
It can be enabled in the future if it's patched in the upstream.